Update Script - Split Into Multiple Files
Split monolithic code into multiple files
