+++ Scripts/gui.py
@@ -2,9 +2,10 @@
 # This file was generated by the Tkinter Designer by Parth Jadhav
 # https://github.com/ParthJadhav/Tkinter-Designer
 
-import YouTubeSpammerPurge as myFunctions
+#import YouTubeSpammerPurge as myFunctions
 
 from pathlib import Path
+from Scripts.utils import make_char_set
 import os
 import sys
 
@@ -118,7 +119,7 @@ def take_input_gui(mode, stripLettersNumbers=False, stripKeyboardSpecialChars=Fa
 
         if mode == "chars":
             # Convert characters string to 'set' of characters
-            returnText = myFunctions.make_char_set(text, stripLettersNumbers=stripLettersNumbers, stripKeyboardSpecialChars=stripKeyboardSpecialChars, stripPunctuation=stripPunctuation)
+            returnText = make_char_set(text, stripLettersNumbers=stripLettersNumbers, stripKeyboardSpecialChars=stripKeyboardSpecialChars, stripPunctuation=stripPunctuation)
 
         outputTextBox.config(state='normal')
         outputTextBox.delete('1.0', END)
