+++ Scripts/shared_imports.py
@@ -1,0 +1,8 @@
+import os
+import sys
+import traceback
+import re
+
+from colorama import init, Fore as F, Back as B, Style as S
+
+__all__ = ['os', 'sys', 're', 'traceback','F', 'B', 'S', 'init']
